requirements.txt
python-telegram-bot==20.3
Pillow==10.0.0
requests==2.31.0